+++ app/assets/resident-room.jpg
