+++ app/blog/big-families/page.mdx
@@ -1,5 +1,5 @@
 import { BlogHeader, BlogImage } from "@/components/blog"
-import shoes from './images/shoes.jpeg';
+import shoes from './assets/shoes.jpeg';
 
 export const metadata = {
   title: 'Big Families',
