+++ app/blog/first-annual-gateway-home-memorial-service/page.mdx
@@ -1,8 +1,8 @@
-import candle from './images/candle.jpeg';
-import events from './images/events.jpeg';
-import pie from './images/pie.jpeg';
-import table from './images/table.jpeg';
-import walkway from './images/walkway.jpeg';
+import candle from './assets/candle.jpeg';
+import events from './assets/events.jpeg';
+import pie from './assets/pie.jpeg';
+import table from './assets/table.jpeg';
+import walkway from './assets/walkway.jpeg';
 import { BlogHeader, BlogImage } from "@/components/blog"
 
 export const metadata = {
