+++ app/blog/gateway-home-dedicates-brick-walkways/page.mdx
@@ -1,8 +1,8 @@
 import { BlogHeader, BlogImage } from "@/components/blog"
-import dougSara from './images/doug-sara.jpg'
-import dougSpeaking from './images/doug-speaking.jpg'
-import sara from './images/sara.jpg'
-import saraSpeaking from './images/sara-speaking.jpg'
+import dougSara from './assets/doug-sara.jpg'
+import dougSpeaking from './assets/doug-speaking.jpg'
+import sara from './assets/sara.jpg'
+import saraSpeaking from './assets/sara-speaking.jpg'
 
 export const metadata = {
   title: 'Gateway Home Dedicates Brick Walkways',
