+++ app/blog/new-patio-and-pergola/page.mdx
@@ -1,8 +1,8 @@
 import { BlogHeader, BlogImage } from "@/components/blog"
-import final from './images/final.jpg'
-import pouring from './images/pouring.jpg'
-import roof from './images/roof.jpg'
-import truck from './images/truck.jpg'
+import final from './assets/final.jpg'
+import pouring from './assets/pouring.jpg'
+import roof from './assets/roof.jpg'
+import truck from './assets/truck.jpg'
 
 export const metadata = {
   title: 'New Patio and Pergola',
