+++ app/volunteer/page.tsx
@@ -3,7 +3,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import Image from "next/image";
-import concert from "./images/concert.jpg";
+import concert from "./assets/concert.jpg";
 
 export default function VolunteerPage() {
   const volunteerOpportunities = [
